# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QMS API Documentation'
  description: 'Secure messenger API with end-to-end encryption'
  contact:
    name: Heruvim6w
    email: heruvim.6w@gmail.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'QMS API Server'
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
paths:
...
  '/api/v1/attachments/{id}':
    get:
      tags:
        - Attachments
      summary: 'Get attachment details'
      description: 'Retrieve metadata about a file attachment including file size, MIME type, and original filename. Only accessible to members of the chat containing the message.'
      operationId: getAttachment
      parameters:
        -
          name: id
          in: path
          description: 'Attachment ID'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'Attachment metadata'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
        403:
          description: 'Access denied - not a member of the chat'
        404:
          description: 'Attachment not found'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Attachments
      summary: 'Delete an attachment'
      description: 'Delete a file attachment from a message. Only the message author can delete attachments. File is removed from storage. Message itself remains.'
      operationId: deleteAttachment
      parameters:
        -
          name: id
          in: path
          description: 'Attachment ID'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'Attachment deleted successfully'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: deleted
                type: object
        403:
          description: 'Access denied - only message author can delete attachments'
        404:
          description: 'Attachment not found'
      security:
        -
          bearerAuth: []
  '/api/v1/attachments/{id}/download':
    get:
      tags:
        - Attachments
      summary: 'Download an attachment file'
      description: 'Download the actual file content of an attachment. File is streamed directly from storage. Content-Type is set based on MIME type.'
      operationId: downloadAttachment
      parameters:
        -
          name: id
          in: path
          description: 'Attachment ID'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'File download stream'
          content:
            application/octet-stream: []
        403:
          description: 'Access denied - not a member of the chat'
        404:
          description: 'Attachment or file not found on server'
      security:
        -
          bearerAuth: []
  '/api/v1/messages/{messageId}/attachments':
    get:
      tags:
        - Attachments
      summary: 'Get all attachments for a message'
      operationId: a3dbafce9d8574173c3f20f4037578b8
      parameters:
        -
          name: messageId
          in: path
          description: 'Message ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'List of attachments for the message'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attachment'
        403:
          description: 'Access denied'
        404:
          description: 'Message not found'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    Attachment:
      title: Attachment
      description: 'Вложение к сообщению (файл, изображение и т. п.)'
      required:
        - message_id
        - file_path
      properties:
        id:
          description: 'Уникальный идентификатор вложения'
          type: integer
          format: int64
          example: 123
        message_id:
          description: 'ID сообщения, к которому прикреплено вложение'
          type: integer
          format: int64
          example: 456
        file_path:
          description: 'Путь к файлу на сервере'
          type: string
          example: /storage/uploads/photo.jpg
        mime_type:
          description: 'MIME‑тип файла (например, image/jpeg)'
          type: string
          example: image/jpeg
          nullable: true
        size:
          description: 'Размер файла в байтах'
          type: integer
          example: 102400
          nullable: true
        name:
          description: 'Оригинальное имя файла (как было загружено)'
          type: string
          example: отчёт_2025.pdf
          nullable: true
        created_at:
          description: 'Дата создания записи'
          type: string
          format: date-time
          example: '2025-01-15T10:30:00Z'
          nullable: true
      type: object
...
    AttachmentResponse:
      title: Attachment
      description: 'File attachment'
      properties:
        id:
          type: integer
        message_id:
          type: integer
        file_path:
          type: string
        mime_type:
          type: string
          example: image/jpeg
        size:
          description: 'File size in bytes'
          type: integer
        name:
          description: 'Original filename'
          type: string
        created_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []